Move author tests to new file
- simple move to move two top level tests to new file
